+++ rest/src/main/java/com/pathz/boot/rest/entity/Order.java
@@ -19,7 +19,7 @@ public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer integer;
+    private Long integer;
 
     @Column(name = "price")
     private Double price;
